Object: Fix ToggleRef disposing when we're being finalized
We should not try to dispose our ToggleRef when Dispose(bool) has been
called by the finalizer, because we should not reference any managed
objects.

Set and check the disposed flag inside of Dispose(bool) so that it's
also done when the object is finalized.

Also clear up a confusion between a local ToggleRef variable in Dispose
and the ToggleRef member field.