Backported the *.csproj / *.sln build process from the "master" codeline to the "2.12" codeline.
This way, a convenient browsing of the sources etc. is possible.

This build has been verified to work properly in
- MonoDevelop 3.0.3.2 on Ubuntu
- Visual Studio 2010 on Win32
- Visual Studio 2012 on Win32

The build produces some warnings, but these have NOT been suppressed,
because they originate from errors/glitches in the generated *.cs files and should be cured.
It's no use hiding them.