MSBuild: fix the solution build (except audit and sample projects)
This commit makes it possible to build any project of the gtk-sharp.sln
from an IDE (except audit and sample projects, which require a bit more
work). This doesn't mean that AUTOTOOLS is deprecated, but just that it
is more comfortable to use an IDE when working on gtk-sharp because it
will offer better auto-completion, and will stop highlight misleading
semantic errors, from now on.