Revert "gtk: Mark the Cairo.Context in Widget.Drawn as owned"
This reverts commit 21bfaa7a9d84218e01529a7b9af28085f6e68306.

After fixing the memory leak in a9a67ab41006b88297f538cc96d2834f8c7e5e89,
native crashes would happen when using a Gtk#-based app, and the
reason is because this commit apparently tried to fix the leak
with the wrong approach of marking the CairoContext as owned.

Stephan Sundermann has confirmed me that the CairoContext passed
in the Draw signal is not marked as transfer-ownership=full in
GObject-Introspection metadata. So unmarking this as owned fixes
the double-free that was causing the crash.