Value: First try to invoke a constructor with IntPtr, bool
In Cairo there the default constructor does not take a ref to the native object making the object invalid when trying to access it. This commit changes the behaviour to search for another constructor that takes an owner variable which is set to false when invoked to indicate that the managed side should take a ref.