Update GTK+ to version 3.14.15
Based on the patches from Mikkel Kruse Johnsen. See the PR:
https://github.com/mono/gtk-sharp/pull/128
While Mikkel has patches on his fork to update step by step
to GTK 3.20 here I am focusing on updating to 3.14.15 and getting
things to the minimal to just compile. Then the idea would be
to fix the possible issues by fixing the metadata.

See that the CSS demo is currently disabled since the CssProvider
is not properly picked up for some reason. We will need to research
a bit on this but for now it is good enough doing it without.

There are also 3 new patches that need to be applied due to the
scanner hanging or giving an error when parsing the source code.
Two of those patches are already in GTK upstream see:
https://git.gnome.org/browse/gtk+/commit/?id=64eca0bb2dad94164d38a89797bf7a439a321b48
https://git.gnome.org/browse/gtk+/commit/?id=fab70d34d443d25c8c68fea30ae7b483749b5e40
The third patch has not yet a workaround that we can merge, it seems
The third patch has not yet a workaround that we can merge, it seems
the ifdef about the GI_SCANNER is giving some issues. We should probably
improve the scanner for this.

See that there are also some generator problems with GtkWidgetClass.
The part of the patch was taken from:
https://github.com/openmedicus/gtk-sharp/commit/5fc657a39cd48ca595992f6bdcbf7bea2c348aed
This workaround fixes the problem below. We should probably research here why this happens and how we can fix it in a better way:
./Widget.cs(111,4): error CS0246: The type or namespace name `GtkWidgetClass' could not be found. Are you missing an assembly reference?
./Widget.cs(112,4): error CS0841: A local variable `klass' cannot be used before it is declared
./Widget.cs(113,4): error CS0103: The name `OverrideClassStruct' does not exist in the current context
./Widget.cs(113,32): error CS0841: A local variable `klass' cannot be used before it is declared