[Perf] Optimize number of allocations and iterations on glib lists.
In the usual case, we'd have to iterate the list every time we query
the Count, so in the for loops, we have O(n^2) complexity, when it
should be queried in O(n).

Made the code use wrapper methods which finish the operation in O(n),
one iteration in native, one in managed+native.