[Gen] Fix leaky implementation of array parameters.
In the case of ArrayParameter, a parameter was only freed if it were an
IManualMarshaller, which is not a lot of the cases, therefore we were
leaking all the string pinvokes.

Only Atk.Relation's constructor and GTypes seem to be faulty and should
not be freed after use, so hardcode those in and do the right thing for
other types.