[GLib] Fix GObject finalizers to finally work.
Also remove useless IsAlive code.

Finalizers didn't work because on owned objects, we didn't replace the
creation ref with the toggle_ref itself. Therefore, we would have to
always dispose every GLib.Object due to this bug.

The problem that was stopping this from happening earlier was the wrong
binding on Gdk.PixbufLoader, which caused a crash, because nobody was
ref-ing the Gdk.Pixbuf, causing memory corruptions.

This properly marks items which exist only in managed to be weak
references, so the toggleref is an actual proper toggleref
implementation