[GLib] Fix GLib.Value leak for passing structs marked with [GTypeStruct]
This would allocate memory for the struct and never free it. Used a ManagedValue wrapper and change the type of the gvalue.

When using an object that has a GType.Pointer type, but is not an IWrapper
or a Pointer, the old bit of could would allocate unmanaged memory via the
runtime, marshal the managed memory to it, then pass it to native.

g_value_set_pointer has no free-ing mechanism, so it's up to the creator
of the GValue to free the memory. When we would dispose the GValue, we
would end up just unsetting the pointer, the memory being leaked forever.

We now allocate a GCHandle and wrap the object being passed. This changes
the return type of the GValue, but at least it does not leak.

The case where would this happen would be when the GValue constructor
being used was the IntPtr variant, or the object variant with a
GType.Pointer type.

We now correctly marshal it back to managed.