[Gtk] Make TreeView go the other way around.
This fixes a bug where we'd unref a TreeModel into oblivion by taking out a model for sorting/populating then placing it back.
Move the mechanism from unref-ing on the setter to ref-ing on the Gtk.TreeView constructor which takes a TreeModel, so it matches the behaviour of the former.