[Gen] Optimize marshalling on some struct pinvokes
This changes native signatures to use by-ref passing instead of having to marshal structs.
The idea behind this is to avoid one extra allocation + free, box operation from StructureToPtr and reconstruction of the value after an unmanaged call
MarshalReturnType from VirtualMethod was using the wrong delegation target.
Also, now that the pango_glyph_item_free generated method has a different signature, bind a manual one so we can build and use it.