[Perf] Optimize object creation by using a faster allocator
This code uses a whole lot of tricks to optimize opaque/object creation.
Create a compiled expression for each type that gets created inside the application. As a downside, we have residual memory for app-lifetime, at the cost of performance.
Using this version, we workaround the slowness of Activator.CreateInstance by using a non-boxing, fully qualified activator delegate that we cache and successfully use to create objects. We get rid of small array creation, IntPtr boxing on every object creation.