[GLib] Optimize ToggleRef freeing.
Given that a ToggleRef can only be freed when:
a) it is explicitly disposed, thus can't be accessed anymore (or it
woudl be an user error)
b) the containing object is finalized, thus can't be accessed anymore.

And that the toggleref is always accessed from the UI thread, we can
conclude that it's safe to free the signal array without copying.

Another main issue in this code is that freeing a toggleref was
accessing the Signals property which creates a new hashtable on access
if it's not created.