[Gdk] Fix pixbuf leak
This forced a pixbuf ref on querying. This was not good, as the managed wrapper already takes a ref. So we'd continously ref the pixbuf on the same object, without setting the Owned property. While we could grab ownership of the pixbuf, better rely on the ref/unref semantics to properly handle this.

Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=51535