[GLib, Gen] Add IgnoreRegistrationAttribute
This attribute allows for a fast path of registration of a gtype. We selectively choose what to register if we mark the type with the attribute. Thus, if you mark something with IgnoreRegistration(DefaultHandlers=false), it'll only register DefaultHandlers for a method, so we don't reflect all the methods of each type when we register the gtype.