Replace strings with const strings (#220)
* Replace strings with const strings

This centralized a large chunk of the hard coded strings. The idea is
to make it easier to convert from raw to gir. For example, RAW uses
"enum", but GIR uses "enumeration". Using the const string we can find
all places in the generator that references "enum" as well as update the
name in only one place.

* Make constants class internal

* Add Constants.cs to automake

* Fix typo in DisableDefaultConstructor

* Fix typo