Fix GLib.Log.Write() format string vulnerability via new glue (#226)
* [GLib] Fix Log.Write() format string vulnerability via new glue

The third parameter of g_logv(), as used in GLib.Log.Write(), is
actually not a message (as advertised in the prototype) but a
format string.

The formatting done via String.Format() may leave format specifiers
like %d unescaped, which can then lead to a crash / SIGSEGV while
executing native code. Note that such format specifiers may get
injected via something that then is supposed to get logged via the
affected GLib.Log.Write().

As I'm told that marshalling to a variable arguments function is
not supported, a new glue function, glibsharp_log_msg(), gets
introduced (in a new glue file, glib/glue/log.c). Its only purpose is
to replace passing the message as format string (which is wrong)
with passing "%s" as format string with the message following as
argument which is used as plain (non-format) string.

* Simplify code by using cdecl pinvoke specifics